Use u32 for VarNode indices and sizes instead of usize
Change VarNode fields space_index and size to u32 to reduce memory usage
and ensure consistency. Update related code to cast usize values as
needed.

+++ jingle/src/analysis/valuation/simple/mod.rs
@@ -240,7 +240,7 @@ impl SimpleValuationState {
                     for (vn, _) in new_state.valuation.direct_writes.items() {
                         let keep = self
                             .arch_info
-                            .get_space(vn.space_index)
+                            .get_space(vn.space_index as usize)
                             .map(|space| space._type != SpaceType::IPTR_INTERNAL)
                             .unwrap_or(true);
                         if !keep {
@@ -258,7 +258,7 @@ impl SimpleValuationState {
                 for (vn, _) in new_state.valuation.direct_writes.items() {
                     let keep = self
                         .arch_info
-                        .get_space(vn.space_index)
+                        .get_space(vn.space_index as usize)
                         .map(|space| space._type != SpaceType::IPTR_INTERNAL)
                         .unwrap_or(true);
                     if !keep {
